requirements.txt
requests == 2.32.4
google-api-python-client == 2.171.0
flask == 3.1.1
openai == 1.92.2
gunicorn 

curl -X POST http://localhost:5000/ \  -H "Content-Type: application/json" \ -d "{\"topic\": \"How AI Is Changing Healthcare\"}" 